refactor(visualization): move file-ingestion parser into loadFile feature
fileParser + fileValidator are consumed only by loadFile in production, so
relocate them to features/loadFile/util/. The one cross-feature consumer (the
globalSettings spec using getNameDataPair) now imports it via loadFile's public
facade, which is the depcruise-legal way to cross features. generatedSchema.json
stays in util/ (generated artifact) and is imported feature->util. mockHelper
already removed. dependency-cruiser 0 errors, 74 tests pass.
