+++ visualization/app/codeCharta/features/globalSettings/components/globalConfigurationDialog/confirmResetMapDialog/confirmResetMapDialog.spec.ts
@@ -5,7 +5,7 @@ import { MockStore, provideMockStore } from "@ngrx/store/testing"
 import { render, screen } from "@testing-library/angular"
 import userEvent from "@testing-library/user-event"
 import "fake-indexeddb/auto"
-import { getNameDataPair } from "../../../../../util/fileParser"
+import { getNameDataPair } from "../../../../loadFile/facade"
 import { LoadFileService } from "../../../../../features/loadFile/facade"
 import { LoadInitialFileService, sampleFile1, sampleFile2 } from "../../../../../features/loadFile/facade"
 import { UrlExtractor } from "../../../../../util/urlExtractor"
