+++ visualization/app/codeCharta/features/loadFile/facade.ts
@@ -1,2 +1,3 @@
 export { LoadFileService } from "./services/loadFile.service"
 export { LoadInitialFileService, sampleFile1, sampleFile2 } from "./services/loadInitialFile.service"
+export { getNameDataPair } from "./util/fileParser"
