+++ visualization/app/codeCharta/features/loadFile/services/loadFile.service.ts
@@ -1,12 +1,12 @@
 import { Injectable, OnDestroy } from "@angular/core"
 import { tap } from "rxjs"
 import { clone } from "../../../util/clone"
-import { CCFileValidationResult } from "../../../util/fileValidator"
+import { CCFileValidationResult } from "../util/fileValidator"
 import { FileState } from "../../../model/files/files"
 import { NameDataPair } from "../../../codeCharta.model"
 import { ErrorDialogService } from "../../../features/shared/components/errorDialog/errorDialog.service"
 import { loadFilesValidationToErrorDialog } from "./loadFilesValidationToErrorDialog"
-import { enrichFileStatesAndRecentFilesWithValidationResults } from "../../../util/fileParser"
+import { enrichFileStatesAndRecentFilesWithValidationResults } from "../util/fileParser"
 import { fileRoot } from "../../../util/fileRoot"
 import { LoadFileStore } from "../stores/loadFile.store"
