+++ visualization/app/codeCharta/features/loadFile/services/loadFilesValidationToErrorDialog.ts
@@ -1,4 +1,4 @@
-import { CCFileValidationResult } from "../../../util/fileValidator"
+import { CCFileValidationResult } from "../util/fileValidator"
 
 export function loadFilesValidationToErrorDialog(fileValidationResults: CCFileValidationResult[]) {
     const htmlMessages = [...buildErrorMessages(fileValidationResults), ...buildWarningsMessages(fileValidationResults)]
