+++ visualization/app/codeCharta/features/loadFile/services/loadInitialFile.service.ts
@@ -10,7 +10,7 @@ import { FileState } from "../../../model/files/files"
 import { MetricQueryParemter } from "../../../state/effects/updateQueryParameters/metricQueryParameter"
 import { ErrorDialogService } from "../../../features/shared/components/errorDialog/errorDialog.service"
 import { buildHtmlMessage } from "./loadFilesValidationToErrorDialog"
-import { getNameDataPair } from "../../../util/fileParser"
+import { getNameDataPair } from "../util/fileParser"
 import { LoadFileService, NO_FILES_LOADED_ERROR_MESSAGE } from "./loadFile.service"
 import { UrlExtractor } from "../../../util/urlExtractor"
 import { LoadInitialFileStore } from "../stores/loadInitialFile.store"
