+++ visualization/app/codeCharta/features/loadFile/util/fileValidator.spec.ts
@@ -4,13 +4,13 @@ import {
     TEST_FILE_CONTENT_INVALID_MAJOR_API,
     TEST_FILE_CONTENT_INVALID_MINOR_API,
     TEST_FILE_CONTENT_NO_API
-} from "../mocks/dataMocks"
-import { CodeMapNode, NameDataPair, NodeType } from "../codeCharta.model"
-import packageJson from "../../../package.json"
+} from "../../../mocks/dataMocks"
+import { CodeMapNode, NameDataPair, NodeType } from "../../../codeCharta.model"
+import packageJson from "../../../../../package.json"
 import { checkErrors, checkWarnings, ERROR_MESSAGES } from "./fileValidator"
-import { fileWithFixedFolders, fileWithFixedOverlappingSubFolders } from "../resources/fixed-folders/fixed-folders-example"
-import { APIVersions, ExportCCFile } from "../codeCharta.api.model"
-import { clone } from "./clone"
+import { fileWithFixedFolders, fileWithFixedOverlappingSubFolders } from "../../../resources/fixed-folders/fixed-folders-example"
+import { APIVersions, ExportCCFile } from "../../../codeCharta.api.model"
+import { clone } from "../../../util/clone"
 
 describe("FileValidator", () => {
     let file: ExportCCFile
