+++ visualization/app/codeCharta/features/navBar/components/navBarFolderButton/navBarFolderButton.e2e.ts
@@ -2,7 +2,7 @@ import { test, expect } from "@playwright/test"
 import { clearIndexedDB, goto } from "../../../../../playwright.helper"
 import { NavBarFolderButtonPageObject } from "./navBarFolderButton.po"
 import { MapSelectorPageObject } from "../mapSelector/mapSelector.po"
-import { ERROR_MESSAGES } from "../../../../util/fileValidator"
+import { ERROR_MESSAGES } from "../../../../loadFile/util/fileValidator"
 import { DialogErrorPageObject } from "../../../../features/shared/components/errorDialog/errorDialog.component.po"
 
 test.describe("NavBarFolderButton", () => {
